fix: update test names after code change
