+++ packages/shared/src/index.ts
@@ -7,7 +7,7 @@ import type {
 
 export { WorkerWSEventEmitter } from './emitter'
 export { createWorkerRPC } from './rpc'
-export { assert, limitConcurrency, normalizeDriveLetter } from './utils'
+export { assert, createQueuedHandler, limitConcurrency, normalizeDriveLetter } from './utils'
 
 export type ExtensionTestSpecification = [
   project: string,
